requirements.txt
numpy>=1.21.0,<2.0.0
scipy>=1.7.0
librosa>=0.8.0
soundfile>=0.10.0
opencv-python==4.10.0.84
moviepy>=1.0.3
tqdm>=4.60.0
streamlit>=1.24.0
pyyaml>=5.4.0
matplotlib>=3.3.0
watchdog>=2.1.0